Move Evento CRUD to agenda app and remove eventos
