+++ dashboard/views.py
@@ -12,7 +12,7 @@ from core.permissions import (
 from organizacoes.models import Organizacao
 from nucleos.models import Nucleo
 from empresas.models import Empresa
-from eventos.models import Evento
+from agenda.models import Evento
 
 User = get_user_model()
