--- eventos/templates/eventos/calendar.html
@@ -1,87 +1,0 @@
-{% extends 'base.html' %}
-{% load eventos_extras static %}
-{% block title %}Calendário | Hubx{% endblock %}
-
-{% block extra_css %}
-{{ block.super }}
-<link rel="stylesheet" href="{% static 'eventos/css/eventos.css' %}">
-{% endblock %}
-
-{% block content %}
-<section class="max-w-6xl mx-auto px-4 py-10">
-  <!-- Cabeçalho -->
-  <div class="mb-6 flex items-center justify-between">
-    <h2 class="text-2xl font-bold text-neutral-900">Calendário de Eventos</h2>
-    <div class="flex items-center gap-4 text-sm text-neutral-700">
-      <a href="?mes={{ mes|add:'-1' }}&ano={{ ano }}"
-         class="px-3 py-1 rounded-xl border border-neutral-300 hover:bg-neutral-100">&lt;</a>
-      <span class="font-medium">{{ mes }}/{{ ano }}</span>
-      <a href="?mes={{ mes|add:'1' }}&ano={{ ano }}"
-         class="px-3 py-1 rounded-xl border border-neutral-300 hover:bg-neutral-100">&gt;</a>
-    </div>
-  </div>
-
-  <!-- Tabela do calendário -->
-  <div class="overflow-x-auto">
-    <table class="w-full text-center border-collapse">
-      <thead>
-        <tr class="bg-neutral-50 text-sm text-neutral-600">
-          {% for weekday in weekdays %}
-            <th class="py-2 px-1 border border-neutral-200">{{ weekday }}</th>
-          {% endfor %}
-        </tr>
-      </thead>
-      <tbody>
-        {% for dia in dias %}
-          {% if forloop.first or dia.weekday == 0 %}<tr>{% endif %}
-          <td class="align-top h-24 border border-neutral-200 text-sm p-1
-                     {% if dia.month != mes %}bg-neutral-50 text-neutral-400{% endif %}
-                     {% if dia in eventos_por_dia %}bg-primary-50{% endif %}
-                     {% if selected_date == dia %}ring-2 ring-primary-500{% endif %}">
-            <a href="?mes={{ dia.month }}&ano={{ dia.year }}&dia={{ dia|date:'Y-m-d' }}#d{{ dia.day }}"
-               class="inline-block text-xs font-semibold text-neutral-800">{{ dia.day }}</a>
-            {% if dia in eventos_por_dia %}
-              <ul class="mt-1 space-y-1 text-xs text-neutral-600">
-                {% for e in eventos_por_dia|dict_get:dia %}
-                  <li id="d{{ dia.day }}" class="truncate">{{ e.titulo }}</li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </td>
-          {% if dia.weekday == 6 %}</tr>{% endif %}
-        {% endfor %}
-      </tbody>
-    </table>
-  </div>
-
-  <!-- Lista de eventos -->
-  {% if selected_date %}
-    <h3 class="text-lg font-semibold text-neutral-900 mt-10 mb-4">Eventos em {{ selected_date|date:'d/m/Y' }}</h3>
-
-    {% if eventos_dia %}
-      <div class="grid gap-4 sm:grid-cols-2 md:grid-cols-3">
-        {% for evento in eventos_dia %}
-          <div class="border border-neutral-200 rounded-2xl shadow-sm p-4 flex flex-col justify-between bg-white">
-            <div>
-              <h4 class="text-base font-medium text-neutral-800">{{ evento.titulo }}</h4>
-              <p class="text-sm text-neutral-500 mt-1">{{ evento.data_hora|time:"H:i" }}</p>
-            </div>
-            <div class="flex gap-2 mt-4">
-              <a href="{% url 'eventos:detail' evento.pk %}"
-                 class="text-sm text-blue-600 hover:underline">Ver</a>
-              {% if user.tipo_id|stringformat:"s" == "1" or user.tipo_id|stringformat:"s" == "2" or user.tipo_id|stringformat:"s" == "4" %}
-                <a href="{% url 'eventos:update' evento.pk %}"
-                   class="text-sm text-yellow-600 hover:underline">Editar</a>
-                <a href="{% url 'eventos:delete' evento.pk %}"
-                   class="text-sm text-red-600 hover:underline">Remover</a>
-              {% endif %}
-            </div>
-          </div>
-        {% endfor %}
-      </div>
-    {% else %}
-      <p class="text-neutral-500 text-sm mt-4">Nenhum evento neste dia.</p>
-    {% endif %}
-  {% endif %}
-</section>
-{% endblock %}
