--- eventos/templates/eventos/list.html
@@ -1,57 +1,0 @@
-{% extends 'base.html' %}
-{% block title %}Eventos | Hubx{% endblock %}
-
-{% block content %}
-<section class="max-w-6xl mx-auto px-4 py-10">
-
-  <!-- Cabeçalho -->
-  <div class="flex flex-col sm:flex-row sm:items-center sm:justify-between mb-6 gap-4">
-    <div>
-      <h2 class="text-2xl font-bold text-neutral-900">Lista de Eventos</h2>
-    </div>
-    {% if user.tipo_id == 4 or user.tipo_id == 1 %}
-      <a href="{% url 'eventos:create' %}" class="inline-flex items-center gap-2 bg-primary-600 text-white text-sm px-4 py-2 rounded-xl hover:bg-primary-700">
-        <span class="text-xl leading-none">+</span> Novo Evento
-      </a>
-    {% endif %}
-  </div>
-
-  <!-- Filtro -->
-  <form method="get" class="mb-8">
-    {{ form.media }}
-    <div class="flex flex-col sm:flex-row gap-4">
-      {{ form.evento|add_class:"form-input w-full sm:max-w-sm rounded-xl border border-neutral-300 px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-primary-500" }}
-    </div>
-    <noscript>
-      <button type="submit" class="mt-2 bg-neutral-200 text-neutral-800 px-4 py-2 rounded-xl text-sm hover:bg-neutral-300">
-        Buscar
-      </button>
-    </noscript>
-  </form>
-
-  <!-- Lista de Eventos -->
-  {% if object_list %}
-    <div class="grid gap-6 sm:grid-cols-2 md:grid-cols-3">
-      {% for evento in object_list %}
-        <div class="bg-white border border-neutral-200 rounded-2xl shadow-sm p-5 flex flex-col justify-between">
-          <div>
-            <h3 class="text-lg font-semibold text-neutral-900">{{ evento.titulo }}</h3>
-            <p class="text-sm text-neutral-500 mt-1">{{ evento.data_hora|date:"d/m/Y H:i" }}</p>
-          </div>
-          <div class="flex justify-end gap-3 mt-4">
-            <a href="{% url 'eventos:detail' evento.id %}" class="text-sm text-blue-600 hover:underline">Ver</a>
-            {% if user.tipo_id == 4 or user.tipo_id == 1 %}
-              <a href="{% url 'eventos:update' evento.id %}" class="text-sm text-yellow-600 hover:underline">Editar</a>
-              <a href="{% url 'eventos:delete' evento.id %}" class="text-sm text-red-600 hover:underline">Remover</a>
-            {% endif %}
-          </div>
-        </div>
-      {% endfor %}
-    </div>
-  {% else %}
-    <div class="text-center bg-white border border-neutral-200 rounded-2xl p-10 shadow-sm mt-6">
-      <p class="text-neutral-500">Nenhum evento.</p>
-    </div>
-  {% endif %}
-</section>
-{% endblock %}
