+++ templates/base.html
@@ -53,8 +53,8 @@
         <a href="{% url 'empresas:lista' %}" class="flex items-center gap-x-2 hover:text-primary transition">
           <i class="fas fa-building"></i> Empresas
         </a>
-        <a href="{% url 'eventos:calendar' %}" class="flex items-center gap-x-2 hover:text-primary transition">
-          <i class="fas fa-calendar-alt"></i> Eventos
+        <a href="{% url 'agenda:calendario' %}" class="flex items-center gap-x-2 hover:text-primary transition">
+          <i class="fas fa-calendar-alt"></i> Agenda
         </a>
         <a href="{% url 'forum:index' %}" class="flex items-center gap-x-2 hover:text-primary transition">
           <i class="fas fa-comments"></i> Fórum
